requirements.txt
Jinja2==3.1.2
PyYAML==6.0.1
pytest==7.4.3
